+++ game/g_save.c
@@ -258,7 +258,7 @@ void InitGame (void)
 void WriteField1 (FILE *f, field_t *field, byte *base)
 {
 	void		*p;
-	int			len;
+	size_t			len;
 	int			index;
 
 	if (field->flags & (FFL_NOSAVE|FFL_SPAWNTEMP))
@@ -281,7 +281,7 @@ void WriteField1 (FILE *f, field_t *field, byte *base)
 			len = strlen(*(char **)p) + 1;
 		else
 			len = 0;
-		*(int *)p = len;
+		*(int *)p = (int)len;
 		break;
 	case F_EDICT:
 		if ( *(edict_t **)p == NULL)
@@ -331,7 +331,7 @@ void WriteField1 (FILE *f, field_t *field, byte *base)
 
 void WriteField2 (FILE *f, field_t *field, byte *base)
 {
-	int			len;
+	size_t			len;
 	void		*p;
 
 	if (field->flags & (FFL_NOSAVE|FFL_SPAWNTEMP) )
